Updated MovingAverageProcess so that sum value is long now

+++ src/main/java/org/bdcourse/process/MovingAverageProcessv2.java
@@ -15,7 +15,7 @@ public class MovingAverageProcessv2 extends ProcessFunction<Tuple2<String, Integ
     Integer sum=0;
 
     private ValueState<Integer> amountState;
-    private ValueState<Integer> sumState;
+    private ValueState<Long> sumState;
 
 
     public MovingAverageProcessv2(Integer amount, Integer sum){
@@ -29,7 +29,7 @@ public class MovingAverageProcessv2 extends ProcessFunction<Tuple2<String, Integ
         amountState = getRuntimeContext().getState(
                 new ValueStateDescriptor<>("Mystate", Integer.class));
         sumState = getRuntimeContext().getState(
-                new ValueStateDescriptor<>("Mystate2", Integer.class));
+                new ValueStateDescriptor<>("Mystate2", Long.class));
 
     }
 
@@ -37,14 +37,14 @@ public class MovingAverageProcessv2 extends ProcessFunction<Tuple2<String, Integ
     public void processElement(Tuple2<String, Integer> value, Context context, Collector<Tuple3<String, Integer, Double>> out) throws Exception {
         if (amountState.value() == null){
             amountState.update(this.amount);
-            sumState.update(this.sum);
+            sumState.update(this.sum.longValue());
         }
         Double average = ((double)sumState.value())/((double)amountState.value());
         Integer intAverage = average.intValue();
 
         Integer tmp1 = amountState.value() + 1;
         amountState.update(tmp1);
-        Integer tmp2 = sumState.value() + value.f1;
+        Long tmp2 = sumState.value() + value.f1;
         sumState.update(tmp2);
         out.collect(new Tuple3<String, Integer, Double>(value.f0, value.f1, average));
     }
